% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create.R
\name{create}
\alias{create}
\title{Create a data source from a data.frame}
\usage{
create(data, name, description = "")
}
\arguments{
\item{data}{A data.frame from which to create the dataset}

\item{name}{Name to assign the dataset on Domo}

\item{description}{Optional description of the dataset}
}
\description{
Create a data source from a data.frame
}
\examples{
df <- data.frame(matrix(rnorm(20), nrow=10))
DomoR::create(df, name="My Data Source Name", description="My Data Source Description")
}

