% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch}
\alias{fetch}
\title{Fetch a data source.}
\usage{
fetch(id, columns = NULL, use.make.names = FALSE, ...)
}
\arguments{
\item{id}{A data source id (GUID) or an index from a previous list.}

\item{columns}{A vector of column names to return from the Domo datasource. (not case sensitive)}

\item{...}{Additional httr options}
}
\value{
A \code{data.frame} built from the requested Domo data source.
}
\description{
Retrieves a data source by ID (GUID) or by
previous list() index from Domo,
converts it to a data.frame and returns.
}
\examples{
DomoR::init(Sys.getenv('DOMO_BASE_URL'), Sys.getenv('DEVELOPER_TOKEN'))
df <- DomoR::fetch('4826e3fb-cd23-468d-9aff-96bf5b690247')
DomoR::list_ds(limit=10)
df <- DomoR::fetch(1)
df <- DomoR::fetch('4826e3fb-cd23-468d-9aff-96bf5b690247',
  c('accountid', 'lastname', 'startdate'),
  httr::progress())
}
