% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_dataset_disk.R
\name{fetch_to_disk}
\alias{fetch_to_disk}
\title{Fetch a data source to disk.}
\usage{
fetch_to_disk(data_source_id, nrows = NULL, delete.tmp.file = TRUE)
}
\arguments{
\item{data_source_id}{A data source id (GUID)}

\item{nrows}{number of rows to return in dataframe by default returns all rows}

\item{delete.tmp.file}{temp file to save the data by default tmp_file is deleted}
}
\value{
A \code{data.frame} built from the requested Domo data source
}
\description{
Retrieves a data source by GUID, saves to temp file and converts it to a data.frame and returns.
}
\examples{
DomoR::init(Sys.getenv('DOMO_BASE_URL'), Sys.getenv('DEVELOPER_TOKEN'))
df <- DomoR::fetch_to_disk(data_source_id="4826e3fb-cd23-468d-9aff-96bf5b690247", nrows=5, delete.tmp.file=TRUE)
}
