% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DomoR.R
\name{init}
\alias{init}
\title{Initialize DomoR Package}
\usage{
init(customer, token, config = NULL, verbose = FALSE)
}
\arguments{
\item{customer}{The customer ID or base URL (e.g., 'acme' or 'acme.domo.com').}

\item{token}{The developer API token from Domo.}

\item{config}{Optional httr configuration settings.}

\item{verbose}{Logical, enable verbose output.}
}
\description{
Initializes the connection settings for the DomoR package.
}
\examples{
DomoR::init(Sys.getenv('DOMO_BASE_URL'), Sys.getenv('DEVELOPER_TOKEN'))
}
