% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_dataflows.R
\name{list_dataflows}
\alias{list_dataflows}
\title{Determine dataflows inputs and outputs based of dataflow datasetId}
\usage{
list_dataflows(outDataSourceId = NULL, use.dataflows.json = FALSE)
}
\arguments{
\item{outDataSourceId}{A data source id (GUID)}

\item{use.dataflows.json}{TRUE returns pretty json output}
}
\description{
Determine dataflows inputs and outputs based of dataflow datasetId
}
\examples{
DomoR::init(Sys.getenv('DOMO_BASE_URL'), Sys.getenv('DEVELOPER_TOKEN'))
DomoR::list_dataflows(outDataSourceId='d0b246-3b41', use.dataflows.json=FALSE)
}
