% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_dataset_versions.R
\name{list_dataset_versions}
\alias{list_dataset_versions}
\title{Determine dataversions based on dataset Id}
\usage{
list_dataset_versions(id, use.dataversions.json = FALSE)
}
\arguments{
\item{id}{A data source id (GUID)}

\item{use.dataversions.json}{TRUE returns pretty json output}
}
\description{
Determine dataversions based on dataset Id
}
\examples{
DomoR::init(Sys.getenv('DOMO_BASE_URL'), Sys.getenv('DEVELOPER_TOKEN'))
DomoR::list_dataset_versions(id='d0b246-3b41')
}
