% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_ds.R
\name{list_ds}
\alias{list_ds}
\title{List data sources}
\usage{
list_ds(limit = 50, name = NULL, offset = 0, owner_id = NULL,
  display_type = NULL, data_provider = NULL, order_by = NULL)
}
\arguments{
\item{limit}{Maximum number of entries to return.  If limit is less than 1, list all datasets that match criteria.}

\item{name}{Filter by partial dataset name (not case sensitive)}

\item{offset}{Starting index into the result}

\item{owner_id}{Filter by dataset owner ID}

\item{display_type}{Filter results by type.  e.g. 'salesforce'}

\item{data_provider}{Filter results by authorization type.  e.g. 'salesforce'}

\item{order_by}{Order the results by column.  e.g. 'name'}
}
\value{
A \code{data.frame} containing a list of available datasets
}
\description{
List data sources
}
\examples{
\dontrun{
  DomoR::init(Sys.getenv('DOMO_BASE_URL'), Sys.getenv('DEVELOPER_TOKEN'))
  DomoR::list_ds()
  DomoR::list_ds(limit=10, name="sales+results")
}
}
