% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owner.R
\name{owner}
\alias{owner}
\title{Determine my Domo user ID}
\usage{
owner()
}
\description{
Determine my Domo user ID
}
\examples{
\dontrun{
  DomoR::init(Sys.getenv('DOMO_BASE_URL'), Sys.getenv('DEVELOPER_TOKEN'))
  DomoR::owner()
}
}
