% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/replace_ds.R
\name{replace_ds}
\alias{replace_ds}
\title{Replace a data source from a data.frame}
\usage{
replace_ds(id, data, ...)
}
\arguments{
\item{id}{A GUID corresponding to the datasource ID from Domo.}

\item{data}{A data.frame from which to create a data source.}

\item{...}{Additional curl and httr parameters}
}
\description{
Replace a data source from a data.frame
}
\examples{
DomoR::init(Sys.getenv('DOMO_BASE_URL'), Sys.getenv('DEVELOPER_TOKEN'))
df <- data.frame(matrix(rnorm(20), nrow=10))
DomoR::replace_ds(id,df)
}

