% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{schema_domo}
\alias{schema_domo}
\title{Returns list of columns from schema of DOMO dataset}
\usage{
schema_domo(data_source_id)
}
\arguments{
\item{data_source_id}{A GUID corresponding to the datasource ID from Domo.}

\item{data}{A data.frame from which to create a data source.}
}
\value{
list of columns with schema details
}
\description{
Returns list of columns from schema of DOMO dataset
}
\examples{
DomoR::init(Sys.getenv('DOMO_BASE_URL'), Sys.getenv('DEVELOPER_TOKEN'))
df <- DomoR::schema_domo(data_source_id)

}
