% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DomoR.R
\name{with.suffix}
\alias{with.suffix}
\title{Append a suffix to a customer name if missing}
\usage{
\method{with}{suffix}(customer, suffix = ".domo.com")
}
\arguments{
\item{customer}{The customer name or base URL.}

\item{suffix}{The suffix to append (default '.domo.com').}
}
\value{
A character string with the suffix appended if necessary.
}
\description{
Ensures the provided customer name includes the '.domo.com' suffix.
}
