% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DomoR.R
\name{init}
\alias{init}
\title{Initialize Domo package}
\usage{
init(customer, token, config = NULL, verbose = FALSE)
}
\arguments{
\item{customer}{The customer ID or base URL.  e.g.  acme.domo.com  or  acme}

\item{token}{The DEV token required for API access.}

\item{config}{Default is \code{NULL}.}

\item{verbose}{Default is \code{TRUE}.}
}
\description{
Initialize Domo package
}
\examples{
\dontrun{
  DomoR::init(Sys.getenv('DOMO_BASE_URL'), Sys.getenv('DEVELOPER_TOKEN'))
}
}
